# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Blog api documentation.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Fetch all posts.",
                "operationId": "a5157d12defea153331f04463de26227",
                "responses": {
                    "200": {
                        "description": "Fetch all posts"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new Post",
                "operationId": "bc5b11839eb2ea36a6124466f8d0e26b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully."
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Fetch the post with specific id.",
                "operationId": "77c8d21023db7c8069e98fe92e52c185",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the post to be fetched.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post fetched successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post with specified id.",
                "operationId": "e924a645dfa8ba7794d9b4e21cc258c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the post to be updated.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Post updated successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post with specified id.",
                "operationId": "a8da4022be64a9780c3a01583cc86cd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the post to be deleted.",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted successfully."
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post with specified id.",
                "operationId": "2043f13dfc76cc825290a02927775303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the post to be updated.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Post updated successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Post": {
                "title": "Post",
                "properties": {
                    "title": {
                        "description": "Title of the post.",
                        "type": "string"
                    },
                    "featured_image": {
                        "description": "Featured image for the post.",
                        "type": "string",
                        "format": "binary"
                    },
                    "category_id": {
                        "description": "Category Id the post belongs to.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description/content for the post.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}